feat(api): update FontUtil and enhance Player entity API
BREAKING CHANGE:
- FontUtil.createFont() methods now accept float size instead of int
- All usages of createFont() need to update size parameter type

Added:
- Player entity control methods:
  - setSneaking(boolean)
  - isSprinting()
  - setSprinting(boolean)

FontUtil changes address initial design flaws by supporting fractional font
sizes for better typography control. The new Player methods provide proper
sneak/sprint control capabilities.

+++ src/main/java/today/opai/api/interfaces/render/FontUtil.java
@@ -11,7 +11,7 @@ public interface FontUtil {
      * @param size The size of the font.
      * @return A {@link Font} object created from the input stream.
      */
-    Font createFont(InputStream inputStream, int size);
+    Font createFont(InputStream inputStream, float size);
 
     /**
      * Creates a true type font (.ttf) from a byte array.
@@ -20,7 +20,7 @@ public interface FontUtil {
      * @param size The size of the font.
      * @return A {@link Font} object created from the byte array.
      */
-    Font createFont(byte[] data, int size);
+    Font createFont(byte[] data, float size);
 
     /**
      * Creates a true type font (.ttf) from a file.
@@ -29,7 +29,7 @@ public interface FontUtil {
      * @param size The size of the font.
      * @return A {@link Font} object created from the file.
      */
-    Font createFont(File fontFile, int size);
+    Font createFont(File fontFile, float size);
 
     Font getVanillaFont();
